+++ lib/llm/provider.rb
@@ -7,6 +7,7 @@
 # @abstract
 class LLM::Provider
   require "net/http"
+  require_relative "provider/transport"
   require_relative "provider/transport/http"
   require_relative "provider/transport/http/execution"
   include Transport::HTTP::Execution
@@ -27,7 +28,9 @@ class LLM::Provider
   # @param [Boolean] persistent
   #  Whether to use a persistent connection.
   #  Requires the net-http-persistent gem.
-  def initialize(key:, host:, port: 443, timeout: 60, ssl: true, base_path: "", persistent: false)
+  # @param [LLM::Provider::Transport, nil] transport
+  #  Optional transport override used to execute requests.
+  def initialize(key:, host:, port: 443, timeout: 60, ssl: true, base_path: "", persistent: false, transport: nil)
     @key = key
     @host = host
     @port = port
@@ -36,7 +39,7 @@ class LLM::Provider
     @base_path = normalize_base_path(base_path)
     @base_uri = URI("#{ssl ? "https" : "http"}://#{host}:#{port}/")
     @headers = {"User-Agent" => "llm.rb v#{LLM::VERSION}"}
-    @transport = Transport::HTTP.new(host:, port:, timeout:, ssl:, persistent:)
+    @transport = transport || Transport::HTTP.new(host:, port:, timeout:, ssl:, persistent:)
     @monitor = Monitor.new
   end